requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
astroid==2.3.1
astropy==3.2.1
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
decorator==4.4.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.10
django-apscheduler==0.6.0
django-cors-headers==3.10.1
django-heroku==0.3.1
djangorestframework==3.13.0
gunicorn==20.1.0
line-bot-sdk==2.0.1
matplotlib==3.1.1
mysql-connector-python==8.0.26
mysqlclient==2.1.0
numpy==1.16.5
pandas==0.25.1
pandocfilters==1.4.2
pylint==2.4.2
PyMySQL==1.0.2
pyOpenSSL==19.0.0
pyparsing==2.4.2
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
requests==2.22.0
schedule==1.1.0
six==1.12.0
soupsieve==1.9.3
SQLAlchemy==1.3.9
sqlparse==0.4.2
static3==0.7.0
urllib3==1.24.2
wrapt==1.11.2
